# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.CreateAuthProviderRequest:
    properties:
      auth_url:
        example: https://accounts.google.com/o/oauth2/auth
        type: string
      client_id:
        type: string
      client_secret:
        type: string
      name:
        example: google
        type: string
      redirect_uris:
        example:
        - https://myapp.com/callback
        items:
          type: string
        type: array
      revokation_url:
        example: https://oauth2.googleapis.com/revoke
        type: string
      token_url:
        example: https://oauth2.googleapis.com/token
        type: string
    required:
    - auth_url
    - client_id
    - client_secret
    - name
    - redirect_uris
    - revokation_url
    - token_url
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.CreateUserRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: password123
        type: string
      roles:
        example:
        - user
        - admin
        items:
          type: string
        type: array
      username:
        example: john_doe
        type: string
    required:
    - email
    - username
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.OAuthCallbackRequest:
    properties:
      code:
        type: string
      state:
        type: string
    required:
    - code
    type: object
  github_com_ketsuna-org_sovrabase_internal_models_requests.ProjectLoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.ProjectSignupRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
      username:
        example: john_doe
        type: string
    required:
    - email
    - password
    type: object
...
  responses.AnalyticsCapability:
    properties:
      anonymous:
        type: boolean
      enabled:
        type: boolean
      monitors_requests:
        type: boolean
    type: object
  responses.AuthCapability:
    properties:
      enabled:
        type: boolean
      mfa_enabled:
        type: string
      providers:
        items:
          $ref: '#/definitions/responses.AuthProvider'
        type: array
      session_timeout:
        type: integer
    type: object
  responses.AuthProvider:
    properties:
      auth_url:
        type: string
      client_id:
        type: string
      client_secret:
        type: string
      name:
        type: string
      redirect_uris:
        items:
          type: string
        type: array
      revokation_url:
        type: string
      token_url:
        type: string
    type: object
  responses.ComplianceSettings:
    properties:
      hipaa:
        type: boolean
      rgpd:
        type: boolean
    type: object
  responses.CreateUserResponse:
    properties:
      data:
        $ref: '#/definitions/responses.CreateUserResponseData'
      message:
        type: string
    type: object
  responses.CreateUserResponseData:
    properties:
      id:
        type: string
      password:
        description: Only set if password was not provided in request
        type: string
    type: object
  responses.DatabaseCapability:
    properties:
      backup_enabled:
        type: boolean
      enabled:
        type: boolean
      engines:
        items:
          type: string
        type: array
      max_connections:
        type: integer
    type: object
  responses.FunctionsCapability:
    properties:
      enabled:
        type: boolean
      max_execution_time:
        type: integer
      memory_limit_mb:
        type: integer
      runtime:
        type: string
    type: object
  responses.OrgSettings:
    properties:
      compliance:
        $ref: '#/definitions/responses.ComplianceSettings'
      multi_tenant:
        type: boolean
      region:
        type: string
    type: object
  responses.Organisation:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      members_count:
        type: integer
      name:
        type: string
      owner_id:
        type: string
      projects_count:
        type: integer
      settings:
        $ref: '#/definitions/responses.OrgSettings'
      status:
        type: string
      updated_at:
        type: string
    type: object
  responses.Project:
    properties:
      capabilities:
        $ref: '#/definitions/responses.ProjectCapabilities'
      cors:
        items:
          type: string
        type: array
      created_at:
        type: string
      id:
        type: string
      members:
        items:
          $ref: '#/definitions/responses.ProjectMember'
        type: array
      name:
        type: string
      org_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  responses.ProjectCapabilities:
    properties:
      analytics:
        $ref: '#/definitions/responses.AnalyticsCapability'
      auth:
        $ref: '#/definitions/responses.AuthCapability'
      database:
        $ref: '#/definitions/responses.DatabaseCapability'
      functions:
        $ref: '#/definitions/responses.FunctionsCapability'
      realtime:
        $ref: '#/definitions/responses.RealtimeCapability'
      storage:
        $ref: '#/definitions/responses.StorageCapability'
    type: object
  responses.ProjectMember:
    properties:
      role:
        type: string
      user:
        $ref: '#/definitions/responses.User'
    type: object
  responses.RealtimeCapability:
    properties:
      enabled:
        type: boolean
      max_connections:
        type: integer
      protocols:
        items:
          type: string
        type: array
    type: object
  responses.StorageCapability:
    properties:
      enabled:
        type: boolean
      max_size_gb:
        type: integer
      provider:
        type: string
      public_access:
        type: boolean
    type: object
  responses.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      roles:
        items:
          type: string
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the Sovrabase API server.
  title: Sovrabase API
  version: "1.0"
paths:
  /admin/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: User creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CreateUserResponse'
      security:
      - Bearer: []
      summary: Create an User (Need to be authentified)
      tags:
      - Admin
...
  /admin/organizations:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Organisation'
            type: array
      security:
      - Bearer: []
      summary: Return the list of organisation created on the server
      tags:
      - Admin
  /admin/projects:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Project'
            type: array
      security:
      - Bearer: []
      summary: Get a list of all projects created on the Server
      tags:
      - Admin
  /admin/users:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.User'
            type: array
      security:
      - Bearer: []
      summary: Get all registered Users !
      tags:
      - Admin
...
  /project/{id}/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.ProjectLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Login a new user in the project !
      tags:
      - Auth
  /project/{id}/auth/logout:
    post:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Remove a session of the User !
      tags:
      - Auth
  /project/{id}/auth/providers:
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Auth provider data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.CreateAuthProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Create Auth Provider
      tags:
      - Auth
  /project/{id}/auth/providers/{provider}/callback:
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Provider Name
        in: path
        name: provider
        required: true
        type: string
      - description: OAuth callback data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.OAuthCallbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: OAuth Callback
      tags:
      - Auth
  /project/{id}/auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Signup data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.ProjectSignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Signup a new User in the project.
      tags:
      - Auth
...
  /project/{id}/user:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete Project User
      tags:
      - Auth
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get Project User
      tags:
      - Auth
...
swagger: "2.0"